Revert "Moved some util functionality from worker to util"
This reverts commit f2347cb05878adf31c9538be489ef174205be761.
