+++ pulsar-functions/utils/src/main/java/org/apache/pulsar/functions/utils/Utils.java
@@ -22,25 +22,10 @@ import com.google.protobuf.AbstractMessage.Builder;
 import com.google.protobuf.MessageOrBuilder;
 import com.google.protobuf.util.JsonFormat;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.net.ServerSocket;
-import java.net.URI;
-
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.pulsar.client.api.MessageId;
 import org.apache.pulsar.client.impl.MessageIdImpl;
-
-import dlshade.org.apache.zookeeper.KeeperException.Code;
-import org.apache.distributedlog.AppendOnlyStreamWriter;
-import org.apache.distributedlog.DistributedLogConfiguration;
-import org.apache.distributedlog.api.DistributedLogManager;
-import org.apache.distributedlog.api.namespace.Namespace;
-import org.apache.distributedlog.exceptions.ZKException;
-import org.apache.distributedlog.impl.metadata.BKDLConfig;
-import org.apache.distributedlog.metadata.DLMetadata;
-
 
 /**
  * Utils used for runtime.
@@ -76,90 +61,4 @@ public class Utils {
         JsonFormat.parser().merge(json, builder);
     }
 
-    public static int findAvailablePort() {
-        // The logic here is a little flaky. There is no guarantee that this
-        // port returned will be available later on when the instance starts
-        // TODO(sanjeev):- Fix this
-        try {
-            ServerSocket socket = new ServerSocket(0);
-            int port = socket.getLocalPort();
-            socket.close();
-            return port;
-        } catch (IOException ex){
-            throw new RuntimeException("No free port found", ex);
-        }
-    }
-
-    public static void uploadToBookeeper(Namespace dlogNamespace,
-                                         InputStream uploadedInputStream,
-                                         String destPkgPath)
-            throws IOException {
-
-        // if the dest directory does not exist, create it.
-        if (dlogNamespace.logExists(destPkgPath)) {
-            dlogNamespace.deleteLog(destPkgPath);
-        }
-
-        try (DistributedLogManager dlm = dlogNamespace.openLog(destPkgPath)) {
-            try (AppendOnlyStreamWriter writer = dlm.getAppendOnlyStreamWriter()){
-
-                try (OutputStream out = new DLOutputStream(dlm, writer)) {
-                    int read = 0;
-                    byte[] bytes = new byte[1024];
-                    while ((read = uploadedInputStream.read(bytes)) != -1) {
-                        out.write(bytes, 0, read);
-                    }
-                    out.flush();
-                }
-            }
-        }
-    }
-
-    public static void downloadFromBookkeeper(Namespace namespace,
-                                              OutputStream outputStream,
-                                              String packagePath) throws IOException {
-        DistributedLogManager dlm = namespace.openLog(packagePath);
-        InputStream in = new DLInputStream(dlm);
-        int read = 0;
-        byte[] bytes = new byte[1024];
-        while ((read = in.read(bytes)) != -1) {
-            outputStream.write(bytes, 0, read);
-        }
-        outputStream.flush();
-    }
-
-    public static DistributedLogConfiguration getDlogConf(int numReplicas) {
-        DistributedLogConfiguration conf = new DistributedLogConfiguration()
-                .setWriteLockEnabled(false)
-                .setOutputBufferSize(256 * 1024)                  // 256k
-                .setPeriodicFlushFrequencyMilliSeconds(0)         // disable periodical flush
-                .setImmediateFlushEnabled(false)                  // disable immediate flush
-                .setLogSegmentRollingIntervalMinutes(0)           // disable time-based rolling
-                .setMaxLogSegmentBytes(Long.MAX_VALUE)            // disable size-based rolling
-                .setExplicitTruncationByApplication(true)         // no auto-truncation
-                .setRetentionPeriodHours(Integer.MAX_VALUE)       // long retention
-                .setEnsembleSize(numReplicas)                     // replica settings
-                .setWriteQuorumSize(numReplicas)
-                .setAckQuorumSize(numReplicas)
-                .setUseDaemonThread(true);
-        conf.setProperty("bkc.allowShadedLedgerManagerFactoryClass", true);
-        conf.setProperty("bkc.shadedLedgerManagerFactoryClassPrefix", "dlshade.");
-        return conf;
-    }
-
-    public static URI initializeDlogNamespace(String zkServers, String ledgersRootPath) throws IOException {
-        BKDLConfig dlConfig = new BKDLConfig(zkServers, ledgersRootPath);
-        DLMetadata dlMetadata = DLMetadata.create(dlConfig);
-        URI dlogUri = URI.create(String.format("distributedlog://%s/pulsar/functions", zkServers));
-
-        try {
-            dlMetadata.create(dlogUri);
-        } catch (ZKException e) {
-            if (e.getKeeperExceptionCode() == Code.NODEEXISTS) {
-                return dlogUri;
-            }
-            throw e;
-        }
-        return dlogUri;
-    }
 }
